Adds named Poller items, removes PollItem class
PollItem wasn't actually doing anything useful -- in fact, it was making
the interface crufty and adding a bunch of special cases to Poller.  So
it's been yanked.  The simplification of Poller's interface made adding
named Poller items easy-peasy, so those are in, too.  Tests still don't
pass because of absent documentation.  I need to get on that.