Fix recv_as() with non-blocking flag
If recv() is provided the ZMQ_NOBLOCK flag, it can return and undefined
value. The recv_as() method currently assumes a ZeroMQ::Message object
is always returned, and calls data(), however this is not always the
case. This can cause the following error:

	Can't call method "data" on an undefined value

Return undefined early in recv_as() if the call to recv() returns
undefined.